refactor: split contexts by model
