+++ src/main/java/coztion/springai/core/ai/application/service/AiService.java
@@ -1,0 +1,10 @@
+package coztion.springai.core.ai.application.service;
+
+import coztion.springai.core.ai.application.model.command.AiGenerateCommand;
+import coztion.springai.core.ai.application.model.result.AiGenerateResult;
+import reactor.core.publisher.Flux;
+
+public interface AiService {
+
+    Flux<AiGenerateResult> generate(AiGenerateCommand command);
+}
