+++ src/main/java/coztion/springai/core/gemini/application/model/command/GeminiGenerateCommand.java
@@ -1,0 +1,99 @@
+package coztion.springai.core.gemini.application.model.command;
+
+import com.google.genai.types.Content;
+import com.google.genai.types.GenerateContentConfig;
+import com.google.genai.types.Part;
+import java.util.Base64;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Stream;
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class GeminiGenerateCommand {
+
+    private GeminiPrompt prompt;
+
+    private GeminiConfig config;
+
+    public Content createContent() {
+        return prompt.createContent();
+    }
+
+    public GenerateContentConfig createConfig() {
+        return config.createConfig();
+    }
+
+    public String getModel() {
+        return config.model;
+    }
+
+    public String getLocation() {
+        return config.location;
+    }
+
+    @Getter
+    @Setter
+    public static class GeminiPrompt {
+
+        private String text;
+
+        private List<GeminiFile> files;
+
+        public Content createContent() {
+            Part textPart = Part.builder().text(text).build();
+
+            List<Part> fileParts = Optional.ofNullable(files).orElse(Collections.emptyList()).stream()
+                    .map(GeminiFile::toPart)
+                    .toList();
+
+            List<Part> parts =
+                    Stream.concat(Stream.of(textPart), fileParts.stream()).toList();
+
+            return Content.builder().parts(parts).role("user").build();
+        }
+
+        @Getter
+        @Setter
+        public static class GeminiFile {
+
+            private String name;
+
+            private String mimeType;
+
+            private String data;
+
+            public Part toPart() {
+                return Part.fromBytes(getDataBytes(), mimeType);
+            }
+
+            private byte[] getDataBytes() {
+                return Base64.getDecoder().decode(data);
+            }
+        }
+    }
+
+    @Getter
+    @Setter
+    public static class GeminiConfig {
+
+        private String model;
+
+        private String location;
+
+        private double temperature;
+
+        private int maxOutputTokens;
+
+        public GenerateContentConfig createConfig() {
+            return GenerateContentConfig.builder()
+                    .temperature((float) temperature)
+                    .maxOutputTokens(maxOutputTokens)
+                    .build();
+        }
+    }
+}
