+++ src/main/java/coztion/springai/core/gemini/application/service/GeminiService.java
@@ -1,0 +1,10 @@
+package coztion.springai.core.gemini.application.service;
+
+import coztion.springai.core.gemini.application.model.command.GeminiGenerateCommand;
+import coztion.springai.core.gemini.application.model.result.GeminiGenerateResult;
+import reactor.core.publisher.Flux;
+
+public interface GeminiService {
+
+    Flux<GeminiGenerateResult> generate(GeminiGenerateCommand generate);
+}
