--- src/main/java/coztion/springai/core/vertex/application/service/VertexService.java
@@ -1,10 +1,0 @@
-package coztion.springai.core.vertex.application.service;
-
-import coztion.springai.core.vertex.application.model.command.VertexGenerateCommand;
-import coztion.springai.core.vertex.application.model.result.VertexGenerateResult;
-import reactor.core.publisher.Flux;
-
-public interface VertexService {
-
-    Flux<VertexGenerateResult> generate(VertexGenerateCommand command);
-}
